Finished connecting the API to the project and added style components
All that is left is to fix the date joined so that it looks more like an actual date. For that I need to figure out how to make a custom time stamp convert. Also need to add the reactive classes for when something is not available or if user is not found